msm8974: added input mediator
Brain dumb simple 2 level input event chain.
For clients that want input events to appear in a
certain order e.g. cpu hotplugging should get it
before cpufreq so that input boost is on both online
cpus

This is used in load_stats and rq_stats hotplugging
governor and in smartmax cpu governor but can be used
in other places too.

Change-Id: I3e3c28cffb640ec0f543eea7994a55bfa4182221